updated the stats page to handle different formats on request

+++ db/migrate/20180222185449_sync_pod_batch_identifiers.rb
@@ -2,18 +2,20 @@ class SyncPodBatchIdentifiers < ActiveRecord::Migration
   require 'csv'
   def up
     puts "#{Dir.pwd}"
-    begin
-      @csv = CSV.read("./tmp/cageShelfIdentifiers.csv", headers: false)
-    rescue
-      opened_file = File.open("./tmp/cageShelfIdentifiers.csv", "r:ISO-8859-1:UTF-8")
-      @csv = CSV.parse(opened_file, headers: false)
-    end
-    CageShelf.transaction do
-      @csv.each_with_index do |row, index|
-        unless index == 0
-          bc = row[0]
-          ident = row[1]
-          CageShelf.where(mdpi_barcode: bc).update_all(identifier: ident)
+    if File.exist?("./tmp/cageShelfIdentifiers.csv")
+      begin
+        @csv = CSV.read("./tmp/cageShelfIdentifiers.csv", headers: false)
+      rescue
+        opened_file = File.open("./tmp/cageShelfIdentifiers.csv", "r:ISO-8859-1:UTF-8")
+        @csv = CSV.parse(opened_file, headers: false)
+      end
+      CageShelf.transaction do
+        @csv.each_with_index do |row, index|
+          unless index == 0
+            bc = row[0]
+            ident = row[1]
+            CageShelf.where(mdpi_barcode: bc).update_all(identifier: ident)
+          end
         end
       end
     end
